services/scrum/requirements.txt
fastapi
uvicorn
python-multipart
motor>=3.3.0
openai
python-docx
pypdf
requests
python-dotenv
tiktoken
